fix: move core game logic to be serial
There really isn't much need to be running the game in parallel. It was
causing loads of race conditions (my fault of course) and was generally
more complex than it needs to be. I've made it so that the player adding
is still in parallel (though this could probably also be serial). In
general, the game manager holds both clients and so there really isn't
any need for parallelism, the main bottleneck of the system is the
users, not the server.